requirements.txt
streamlit
transformers
torch
diffusers
pillow
python-dotenv
huggingface_hub
gTTS
